Add cpu metric record persistence, but not test.

+++ apm-collector/apm-collector-stream/src/main/java/org/skywalking/apm/collector/stream/worker/impl/data/Operation.java
@@ -8,7 +8,7 @@ public interface Operation {
 
     Long operate(Long newValue, Long oldValue);
 
-    Float operate(Float newValue, Float oldValue);
+    Double operate(Double newValue, Double oldValue);
 
     Integer operate(Integer newValue, Integer oldValue);
